+++ src/components/Title.tsx
@@ -1,0 +1,7 @@
+interface Props {
+  children: React.ReactNode;
+}
+
+export default function Title({ children }: Props) {
+  return <h1>{children}</h1>;
+}
